Bengali To English Sentence Learning
This is a Java Desktop Based Application, which is used to store
sentences, words with meaning. User can store sentences, words easily
from the settings panel.
The important feature is DB will be automatically updated and if user
want to uninstall this software, the data will not be lost and user can
regain the data by installing the software again.

Front view of the software is given in SOFT IMG folder.